refactor(cli): split reset.go by destructive flow (E5)
Separate the operator identity-state wipe into reset_identity.go and the
agent-account teardown (survey/plan/exec + confirms) into
reset_account.go. reset.go keeps the command and the resetE/resetAll
dispatch. Intra-package moves only.
